Inline ScipShardWalker into ScipSemanticdb.findShards
